Prototype ARM64 EFI build (also Ten64)
This adds support for ARM64 Embedded & Server Boot systems
which use grub2 and EFI to boot. As OpenSuSE also embeds a copy
of U-Boot for certain targets it may be possible to use this
for the Raspberry Pi 3 and 4 with slight adjustments.

This image also include drivers for the Traverse Ten64 board.