add potentially missing dependencies of btrfsprogs and gfxboot #26
In some host instance of the suggested Leap 15.2, i.e. some vagrant
boxes and in the Pi4 images for Leap15.2 there is not default
install/use of btrfsprogs. Fix by adding to instructions as our target
image construction relies absolutely on this packages. As well as the
also found to be missing in some instances gfxboot.