Move default iso/rockstor rpm version to 4.0.7-0 #57
Our latest available testing rpm is now 4.0.7-0 and given it's positive
reports on our forum it now looks ready to be promoted to the default
used in out installer build (pre-edit). Note that our prior use of
4.0.0-0 and in-turn 4.0.4-0 were instrumental in testing our now zypper
based updates as newer rpms were release. The former indicating a
failure initially in updates, and the latter proving the fix.