Update EULA re https, abstract liability to sub amount & redo dates #83
Adds additional filters that were missed previously and updates the
copyright dates to be current and to catch some new Leap 15.3 dates.

Main change here is to remove a hard-coded $ amount re corner case
liability and replace with "1ST SUBSCRIPTION PAYED"; connecting this
amount to a subscription payment: the 1st in this case. It is postulated
that the original 50 dollars specified in our upstream is derived from
the earlier openSUSE box-set cost.