Temp disable 15.4 Rockstor-Testing repo #123
This repos does not yet exist, so disabling
temporarily to enable testing of the proposed rpms
that are intended to be uploaded to start this new
repo off. There-by enabling installer builds to
check if all is as it should be, before uploading
the intended rpms and building the 15.4 installer that
will include them.