Tumbleweed build failures on all profiles #152
Remove to-be-redundant gio-branding-upstream to avoid
current build failure for all Tumbleweed profiles, and
prepare for likely related changes re ALP.

Thought to be invoked only by glib-devel dependency
in rockstor package which in turn is thought to be from
the to-be-replaced python-dbus use.

Add 'diffutils' to avoid TW profile dependency breaks re
otherwise prioritised busybox-diffutils. Which then
blocks snapper install as busybox-diffutils is not
sufficient.

Includes:
- incidental _multibuild profile updates.